# gateway/api-gateway/openapi.yaml
...
openapi: 3.0.3
info:
  title: Food Delivery API
  version: 1.0.0
  description: |
    Every request goes through the **api-gateway**, which routes it to the owning microservice.

    **To call protected endpoints:** run `POST /login` (e.g. `jamie@example.com` / `password123`),
    copy the `token` from the response, click **Authorize** and paste it.
    For the **Owner** endpoints, log in as a restaurant owner instead
    (e.g. `sam@mariospizzeria.com` / `password123`).

    **Order flow (event-driven over RabbitMQ):** `POST /orders` publishes `order.created` →
    payment-service charges it and publishes `payment.succeeded` or `payment.failed` →
    order-service moves the order to `CONFIRMED` or `CANCELLED` → restaurant-service publishes
    `restaurant.order_received`. The flow is asynchronous, so poll `GET /orders/{id}` for the
    outcome. Orders over 100.00 are declined by the simulated payment provider, so the failure
    path can be tested too.

    **Restaurant side:** a paid (`CONFIRMED`) order shows up in `GET /owner/orders`. The owner
    accepts it (`PREPARING`, publishes `order.preparing`), marks it ready (`READY`, publishes
    `order.ready`) or rejects it (`CANCELLED` and refunded). Customer and owner both get every
    status change pushed over the notification websocket.
servers:
  - url: /
    description: This api-gateway
...
paths:
...
  /owner/restaurants:
    get:
      tags: [Owner restaurants]
      summary: List your restaurants, with their menus
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Your restaurants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RestaurantDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Owner restaurants]
      summary: Register a new restaurant
      description: New restaurants start closed; open them with `PATCH` once the menu is set up.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewRestaurant' }
      responses:
        '201':
          description: Created; publishes `restaurant.created`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestaurantDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /owner/restaurants/{id}:
    get:
      tags: [Owner restaurants]
      summary: Get one of your restaurants
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RestaurantIdPath' }
      responses:
        '200':
          description: The restaurant and its menu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestaurantDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Owner restaurants]
      summary: Update details, or open / close the restaurant
      description: Send only the fields to change. Publishes `restaurant.updated`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RestaurantIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestaurantUpdate' }
            example: { isOpen: false }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestaurantDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /owner/restaurants/{id}/menu-items:
    post:
      tags: [Owner restaurants]
      summary: Add a menu item
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RestaurantIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewMenuItem' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /owner/restaurants/{id}/menu-items/{itemId}:
    patch:
      tags: [Owner restaurants]
      summary: Edit a menu item, or mark it (un)available
      description: Send only the fields to change.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RestaurantIdPath' }
        - { $ref: '#/components/parameters/MenuItemIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MenuItemUpdate' }
            example: { isAvailable: false }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Owner restaurants]
      summary: Remove a menu item
      description: Past orders keep their own copy of the item's name and price.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RestaurantIdPath' }
        - { $ref: '#/components/parameters/MenuItemIdPath' }
      responses:
        '204': { description: Removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RestaurantIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, example: 33333333-3333-3333-3333-333333333331 }
    MenuItemIdPath:
      name: itemId
      in: path
      required: true
      schema: { type: string, example: 44444444-4444-4444-4444-444444444441 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Logged in, but not as a restaurant owner
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    RestaurantSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        cuisine: { type: string, nullable: true }
        address: { type: string }
        isOpen: { type: boolean }
    MenuItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: string, example: '12.5' }
        isAvailable: { type: boolean }
    RestaurantDetail:
      allOf:
        - $ref: '#/components/schemas/RestaurantSummary'
        - type: object
          properties:
            description: { type: string, nullable: true }
            menuItems:
              type: array
              items: { $ref: '#/components/schemas/MenuItem' }
    NewRestaurant:
      type: object
      required: [name, address]
      properties:
        name: { type: string, maxLength: 100 }
        address: { type: string, maxLength: 500 }
        cuisine: { type: string, nullable: true, maxLength: 100 }
        description: { type: string, nullable: true, maxLength: 500 }
      example: { name: 'Pasta Corner', address: '3 Canal Road', cuisine: 'Italian' }
    RestaurantUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        address: { type: string, maxLength: 500 }
        cuisine: { type: string, nullable: true, maxLength: 100 }
        description: { type: string, nullable: true, maxLength: 500 }
        isOpen: { type: boolean }
    NewMenuItem:
      type: object
      required: [name, price]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, nullable: true, maxLength: 500 }
        price: { type: number, minimum: 0.01, maximum: 10000, description: 'At most 2 decimals' }
        isAvailable: { type: boolean, default: true }
      example: { name: 'Calzone', description: 'Folded pizza with ham and ricotta', price: 13.5 }
    MenuItemUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, nullable: true, maxLength: 500 }
        price: { type: number, minimum: 0.01, maximum: 10000 }
        isAvailable: { type: boolean }
...